refactor: QueryLog 클래스 패키지 경로 수정
1. QueryLog 클래스를 domain.vo 패키지에서 domain 패키지로 이동한다.
2. 관련된 모든 import 경로를 수정한다.
3. 패키지 구조 변경으로 모듈 간 경계를 명확히 한다.

Resolves #19
