+++ src/main/java/com/posicube/assignment/querylog/adapter/persistence/SpringDataJpaQueryLogRepository.java
@@ -1,6 +1,6 @@
 package com.posicube.assignment.querylog.adapter.persistence;
 
-import com.posicube.assignment.querylog.domain.vo.QueryLog;
+import com.posicube.assignment.querylog.domain.QueryLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SpringDataJpaQueryLogRepository extends JpaRepository<QueryLog, Long> {
