+++ src/main/java/com/posicube/assignment/querylog/domain/port/QueryLogRepository.java
@@ -1,6 +1,6 @@
 package com.posicube.assignment.querylog.domain.port;
 
-import com.posicube.assignment.querylog.domain.vo.QueryLog;
+import com.posicube.assignment.querylog.domain.QueryLog;
 
 public interface QueryLogRepository {
     QueryLog save(QueryLog queryLog);
